# Assignment refactor