--- a/src/compute_offsets.cpp
+++ b/src/compute_offsets.cpp
@@ -32,7 +32,6 @@ void compute_offsets(AST::CallExpression* ast, int frame_offset) {
 }
 
 void compute_offsets(AST::AssignmentExpression* ast, int frame_offset) {
-	frame_offset++; // TODO: delete
 	compute_offsets(ast->m_target, frame_offset++);
 	compute_offsets(ast->m_value, frame_offset++);
 }